Program to find a knight's tour which makes a semimagic square
A program to produce magic knight's tours, as described in the "Knight
Errant" puzzle on page 48 of The Ultimate Book of Number Puzzles by
Kenneth Kelsey.

A magic knight's tour is a knight's tour of an 8x8 chessboard, starting
on any square in the leftmost column of the board and finishing on any
square on the rightmost column, with the following property. The numbers
1, 2, 3, and so on up to 64 are written in turn on each square visited
by the knight, and at the end of the tour, the sum of the numbers in
each row must equal 260, the sum in each column must equal 260, and the
sum midway through each row or column must equal 130.

Kelsey writes "I have been unable to work out more than four variations
of this interesting knight's move square - ignoring rotations and
reflections." Incomplete representations of the four tours found by
Kelsey are presented in the book as puzzles, and the reader is invited
to complete them.

This program finds these four magic knight's tours and four additional
magic knight's tours.

Assuming the tours must start on the leftmost column and finish on the
rightmost column (a condition implied by how the puzzle is framed in the
book, but not explicitly stated), it is sufficient to find all tours
starting in the top four squares of the leftmost column, because the
symmetry of the puzzle's conditions implies that solutions starting in
the bottom four squares of the leftmost column will only be reflections
of those starting in the top four squares.

Therefore, if the code is correct, only eight variations of the
magic knight's tour exist on an 8x8 board, not counting rotations and
reflections.

The development, optimisation and debugging of this code over the past
few days mercifully predates the creation of this Git repository.