Improve compile-time defines and Makefile
The compile-time defines to turn on and off various modes and
optimisations are now 0/1 rather than absent/present. This means they
can be set to 0 or 1 by the Makefile.

The makefile rules ketest5 and ketest6 make a binary with TOUR_TEST
defined which counts the number of knight's tours on a 5x5 or 6x6 board,
respectively.

There is a new #define RESTRICT_START_AND_END. When defined to true, we
keep the existing behaviour where we require the 8x8 semimagic square
solution to start on the left column and end on the right. When defined
to false, the tour can start or end anywhere, although we only actually
search half of one quadrant because any other tours will be rotations
and/or reflections of the tours found from these squares.

The assumption that if a line of four in a quadrant has two numbers
filled in, their sum must be in the range [8, 122] only holds if the 1
and 64 are on opposite sides, so this assumption is not made if
RESTRICT_START_AND_END is false.

The help (-h) now shows the result of the compile-time defines.

Call fflush(stdout) after outputting each tour.

If we emitted any closed knight's tours, we now print the count of
closed tours at the end as well as the total number of tours.